added last_sync block getters and setters to a LastSync struct
