+++ sync-pipelines-rust/src/main.rs
@@ -8,8 +8,8 @@ use crate::db::token_transfers::TokenTransfers;
 use crate::db::transactions::Transactions;
 use crate::db::user_transactions::UserTransaction;
 use crate::settings::Settings;
+use crate::utils::last_sync::LastSync;
 use crate::utils::table::Table;
-use crate::utils::utils::Utils;
 
 pub mod db;
 pub mod settings;
@@ -38,7 +38,7 @@ async fn main() {
 
     match table_name {
         Table::TokenTransfers => {
-            let block_number = Utils::get_last_synced_block_token_transfers();
+            let block_number = LastSync::get_last_synced_block_token_transfers();
             let token_transfers = TokenTransfers::get(pool.clone(), block_number).await;
             if token_transfers.clone().len() == 0 {
                 exit(0);
@@ -46,17 +46,17 @@ async fn main() {
             let number = TokenTransfers::get_max_block_number(token_transfers.clone());
             UserTransaction::insert(pool, UserTransaction::from_token_transfers(token_transfers))
                 .await;
-            Utils::update_last_synced_block_token_transfers(number);
+            LastSync::update_last_synced_block_token_transfers(number);
         }
         Table::Transactions => {
-            let block_number = Utils::get_last_synced_block_transactions();
+            let block_number = LastSync::get_last_synced_block_transactions();
             let transactions = Transactions::get(pool.clone(), block_number).await;
             if transactions.clone().len() == 0 {
                 exit(0);
             }
             let number = Transactions::get_max_block_number(transactions.clone());
             UserTransaction::insert(pool, UserTransaction::from_transactions(transactions)).await;
-            Utils::update_last_synced_block_transactions(number);
+            LastSync::update_last_synced_block_transactions(number);
         }
     };
 }
