+++ sync-pipelines-rust/src/utils/mod.rs
@@ -1,2 +1,3 @@
+pub mod last_sync;
 pub mod table;
 pub mod utils;
